+++ src/themes/classic/galleria.classic.js
@@ -68,6 +68,8 @@ Galleria.addTheme({
                 if ( e.index === options.show ) {
                     $(e.thumbTarget).css('opacity',1);
                 }
+            } else {
+                $(e.thumbTarget).css('opacity', e.index == options.show ? 1 : 0.6);
             }
         });